+++ macrokrs/config.h
@@ -1,63 +1,11 @@
-#ifndef CONFIG_H
-#define CONFIG_H
+#pragma once
 
-#include "config_common.h"
-
-/* USB Device descriptor parameter */
-// #define VENDOR_ID       0xFEED
-// #define PRODUCT_ID      0x6060
-// #define DEVICE_VER      0x0001
-// #define MANUFACTURER    qmkbuilder
-// #define PRODUCT         keyboard
-// #define DESCRIPTION     Keyboard
-
-/* key matrix size */
-#define MATRIX_ROWS 4
-#define MATRIX_COLS 5
-
-/* key matrix pins */
-//#define MATRIX_ROW_PINS { F6, F7, B1, B3 }
-//#define MATRIX_COL_PINS { C6, D7, E6, B4, B5 }
-//#define UNUSED_PINS
-
-/* COL2ROW or ROW2COL */
-//#define DIODE_DIRECTION COL2ROW
-
-/* number of backlight levels */
-
-// #ifdef BACKLIGHT_PIN
-// #define BACKLIGHT_LEVELS 0
-// #endif
-
-/* Set 0 if debouncing isn't needed */
-//#define DEBOUNCING_DELAY 5
-
-/* Mechanical locking support. Use KC_LCAP, KC_LNUM or KC_LSCR instead in keymap */
-#define LOCKING_SUPPORT_ENABLE
-
-/* Locking resynchronize hack */
-#define LOCKING_RESYNC_ENABLE
-
-/* key combination for command */
-#define IS_COMMAND() ( \
-    keyboard_report->mods == (MOD_BIT(KC_LSHIFT) | MOD_BIT(KC_RSHIFT)) \
-)
-
-/* prevent stuck modifiers */
-//#define PREVENT_STUCK_MODIFIERS
-
-
-#ifdef RGB_DI_PIN
-//#define RGBLIGHT_ANIMATIONS
-#define RGBLED_NUM 0
-#define RGBLIGHT_HUE_STEP 8
-#define RGBLIGHT_SAT_STEP 8
-#define RGBLIGHT_VAL_STEP 8
-#endif
-
-
-//#define ENCODERS_PAD_A { B2, D0 }
-//#define ENCODERS_PAD_B { B6, D4 }
-//#define ENCODER_RESOLUTIONS { 2, 4 }
-
-#endif
+#define ANALOG_JOYSTICK_X_AXIS_PIN F4
+#define ANALOG_JOYSTICK_Y_AXIS_PIN F5
+#define ANALOG_JOYSTICK_X2_AXIS_PIN F7
+#define ANALOG_JOYSTICK_Y2_AXIS_PIN F6
+#define ANALOG_JOYSTICK_AXIS_MIN 0
+#define ANALOG_JOYSTICK_AXIS_MAX 1023
+#define ANALOG_JOYSTICK_SPEED_REGULATOR 10
+#define ANALOG_JOYSTICK_READ_INTERVAL 100
+#define ANALOG_JOYSTICK_SPEED_MAX 1
