+++ macrokrs/driver/multi_analog_joystick.h
@@ -19,16 +19,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-#define ANALOG_JOYSTICK_X_AXIS_PIN F4
-#define ANALOG_JOYSTICK_Y_AXIS_PIN F5
-#define ANALOG_JOYSTICK_X2_AXIS_PIN F7
-#define ANALOG_JOYSTICK_Y2_AXIS_PIN F6
-#define ANALOG_JOYSTICK_AXIS_MIN 0
-#define ANALOG_JOYSTICK_AXIS_MAX 1023
-#define ANALOG_JOYSTICK_SPEED_REGULATOR 10
-#define ANALOG_JOYSTICK_READ_INTERVAL 100
-#define ANALOG_JOYSTICK_SPEED_MAX 1
-
 typedef struct {
     int8_t x;
     int8_t y;
