+++ macrokrs/info.json
@@ -3,6 +3,8 @@
     "manufacturer": "CarrasTech",
     "url": "", 
     "maintainer": "CarrasTech",
+    "processor": "atmega32u4",
+    "bootloader": "caterina",
     "usb": {
         "vid": "0x4443",
         "pid": "0x0CBA",
@@ -13,28 +15,34 @@
         "rows": ["B1", "B3", "B2", "B6"]
     },
     "diode_direction": "COL2ROW",
+    "features": {
+        "bootmagic": true,
+        "console": true,
+        "extrakey": true,
+        "nkro": true
+    },
     "layouts": {
         "LAYOUT": {
             "layout": [
-                {"label":"K00", "x":0, "y":0}, 
-                {"label":"K01", "x":1, "y":0}, 
-                {"label":"K02", "x":2, "y":0}, 
-                {"label":"K03", "x":3, "y":0}, 
-                {"label":"K04", "x":4, "y":0},
+                {"label":"K00", "x":0, "y":0, "matrix": [0, 0]},
+                {"label":"K01", "x":1, "y":0, "matrix": [0, 1]},
+                {"label":"K02", "x":2, "y":0, "matrix": [0, 2]},
+                {"label":"K03", "x":3, "y":0, "matrix": [0, 3]},
+                {"label":"K04", "x":4, "y":0, "matrix": [0, 4]},
 
-                {"label":"K10", "x":0, "y":1}, 
-                {"label":"K11", "x":1, "y":1}, 
-                {"label":"K12", "x":2, "y":1}, 
-                {"label":"K13", "x":3, "y":1}, 
-                {"label":"K14", "x":4, "y":1},
+                {"label":"K10", "x":0, "y":1, "matrix": [1, 0]},
+                {"label":"K11", "x":1, "y":1, "matrix": [1, 1]},
+                {"label":"K12", "x":2, "y":1, "matrix": [1, 2]},
+                {"label":"K13", "x":3, "y":1, "matrix": [1, 3]},
+                {"label":"K14", "x":4, "y":1, "matrix": [1, 4]},
 
-                {"label":"K20", "x":0, "y":2}, 
-                {"label":"K21", "x":1, "y":2}, 
-                {"label":"K22", "x":2, "y":2}, 
-                {"label":"K23", "x":3, "y":2},
+                {"label":"K20", "x":0, "y":2, "matrix": [2, 0]},
+                {"label":"K21", "x":1, "y":2, "matrix": [2, 1]},
+                {"label":"K22", "x":2, "y":2, "matrix": [2, 2]},
+                {"label":"K23", "x":3, "y":2, "matrix": [2, 3]},
                  
-                {"label":"K31", "x":1, "y":3}, 
-                {"label":"K32", "x":2, "y":3}
+                {"label":"K31", "x":1, "y":3, "matrix": [3, 1]},
+                {"label":"K32", "x":2, "y":3, "matrix": [3, 2]}
             ]
         }
     }
