--- macrokrs/macrokrs.c
@@ -1,1 +1,0 @@
-#include "macrokrs.h"
